engine/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
geckolib
